FxCop clean up: threading, portability, dead code, unnecessary casting.
git-svn-id: http://svn.uktrainsim.com/svn/openrails/trunk@1767 a9f3cfff-6c41-7446-b0ca-052bd761e6eb

+++ Source/RunActivity/Viewer3D/UserInput.cs
@@ -54,7 +54,7 @@ namespace ORTS
         public static RailDriverState RDState = null;
 
         [DllImport("user32.dll")]
-        static extern int GetAsyncKeyState(Keys key);
+        static extern short GetAsyncKeyState(Keys key);
 
         public static void Update(Viewer3D viewer)
         {
